requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.4.0
certifi==2025.10.5
frozenlist==1.8.0
idna==3.11
magic-filter==1.0.12
multidict==6.7.0
propcache==0.4.1
pydantic==2.11.10
pydantic_core==2.33.2
python-dotenv==1.1.1
typing-inspection==0.4.2
typing_extensions==4.15.0
yarl==1.22.0
httpx>=0.23.0,<1.0.0
